rename starting file to 'main.js'
