+++ package.json
@@ -2,7 +2,7 @@
   "name": "@peter.naydenov/fsm",
   "version": "4.0.2",
   "description": "Simple Finite State Machine",
-  "main": "src/index.js",
+  "main": "src/main.js",
   "type": "module",
   "scripts": {
     "test": "mocha test",
